changed login php sessions variables

+++ questions.php
@@ -6,8 +6,8 @@ Questions Page
 
 <?php
 	// Start a PHP session
-	session_name("admin");
-	session_start("admin");
+	session_name("logged");
+	session_start("logged");
 	
 	// Check to see if user is NOT logged in to prevent unauthorized access
 	/*if (!isset($_SESSION["admin"]))
@@ -39,9 +39,9 @@ Questions Page
 			<span class="left"><a href="menu.php"><img src="images/headerlogo.png" alt="Aeroapps Logo" /></a></span>
 			<?php
 			
-			if (isset($_SESSION["admin"]))
+			if (isset($_SESSION["name"]))
 			{
-				echo '<span class="right">Welcome, ' . $_SESSION["admin"] . '!</span><br />';
+				echo '<span class="right">Welcome, ' . $_SESSION["name"] . '!</span><br />';
 			}
 			else
 			{
@@ -71,21 +71,34 @@ Questions Page
     </div>
     <div id=list>
     	<?php
-			echo "<table width='100%' cellspacing='6px' style='border: solid 1px black;'>";
-			echo "<tr><th width='10%'>ID</th><th>Question Text</th><th>View</th><th>Edit</th><th>Delete</th></tr>";
 
 			include('connect/local-connect.php');
 			
 
 			$query = mysql_query("SELECT * FROM test_questions ORDER BY qID asc") or die(mysql_error($dbc));
 			if(mysql_num_rows($query) > 0) {
-    			while($row = mysql_fetch_array($query)) {
-        			echo "<tr><td style='outline: thin solid black'>".$row['qID']."</td>";
-        			echo "<td style='outline: thin solid black'>".$row['qText']."</td>";
-        			echo "<td style='outline: thin solid black'><form action='viewquestion.php' method='POST'><input type='hidden' name='tempViewID' value='".$row["qID"]."'/><input type='submit' name='view-btn' value='View' /></form></td>";
-					echo "<td style='outline: thin solid black'><form action='editquestion.php' method='POST'><input type='hidden' name='tempEditID' value='".$row["qID"]."'/><input type='submit' name='edit-btn' value='Edit' /></form></td>";
-					echo "<td style='outline: thin solid black'><form action='deletequestionprocess.php' method='POST' onSubmit=\"return confirm('Are you sure you want to delete?')\"><input type='hidden' name='tempDeleteID' value='".$row["qID"]."'/><input type='submit' name='delete-btn' value='Delete' /></form></td></tr>";
+				if ($_SESSION['type'] == 'admin') 
+				{
+					echo "<table width='100%' cellspacing='6px' style='border: solid 1px black;'>";
+					echo "<tr><th width='10%'>ID</th><th>Question Text</th><th>View</th><th>Edit</th><th>Delete</th></tr>";
+    				while($row = mysql_fetch_array($query)) {
+        				echo "<tr><td style='outline: thin solid black'>".$row['qID']."</td>";
+        				echo "<td style='outline: thin solid black'>".$row['qText']."</td>";
+        				echo "<td style='outline: thin solid black'><form action='viewquestion.php' method='POST'><input type='hidden' name='tempViewID' value='".$row["qID"]."'/><input type='submit' name='view-btn' value='View' /></form></td>";
+						echo "<td style='outline: thin solid black'><form action='editquestion.php' method='POST'><input type='hidden' name='tempEditID' value='".$row["qID"]."'/><input type='submit' name='edit-btn' value='Edit' /></form></td>";
+						echo "<td style='outline: thin solid black'><form action='deletequestionprocess.php' method='POST' onSubmit=\"return confirm('Are you sure you want to delete?')\"><input type='hidden' name='tempDeleteID' value='".$row["qID"]."'/><input type='submit' name='delete-btn' value='Delete' /></form></td></tr>";
+    				}
     			}
+    			if ($_SESSION['type'] == 'user')
+    			{
+    				echo "<table width='100%' cellspacing='6px' style='border: solid 1px black;'>";
+					echo "<tr><th width='10%'>ID</th><th>Question Text</th><th>View</th></tr>";
+    				while($row = mysql_fetch_array($query)) {
+        				echo "<tr><td style='outline: thin solid black'>".$row['qID']."</td>";
+        				echo "<td style='outline: thin solid black'>".$row['qText']."</td>";
+        				echo "<td style='outline: thin solid black'><form action='viewquestion.php' method='POST'><input type='hidden' name='tempViewID' value='".$row["qID"]."'/><input type='submit' name='view-btn' value='View' /></form></td>";
+    				}
+    			}	
 			}
 			
 			echo "</table>";
